+++ pliq/__init__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 """PL-Quality (PLIQ): end-to-end docking pose QA (e-DockQ, PoseBusters, TM-score, BINANA)."""
 
-__version__ = "0.6.2"
+__version__ = "0.6.3"
 
 from pliq.e2e import run_pliq_from_pdbs
-from pliq.edockq.binana_recall import DEFAULT_BINANA_OBABEL_PH, DEFAULT_BINANA_STRIP_H
+from pliq.edockq.binana_recall import DEFAULT_BINANA_OBABEL_PH, DEFAULT_BINANA_STRIP_H, resolve_binana_h_settings
 from pliq.edockq.pose_eval import evaluate_pose_from_pdbs
 
 __all__ = [
@@ -13,4 +13,5 @@ __all__ = [
     "evaluate_pose_from_pdbs",
     "DEFAULT_BINANA_OBABEL_PH",
     "DEFAULT_BINANA_STRIP_H",
+    "resolve_binana_h_settings",
 ]